refactor(structure): Migrate to Feature-First project structure
This commit refactors the entire project structure from a file-type-first approach to a more scalable and maintainable feature-first architecture. This change improves modularity and makes the codebase easier to navigate and scale.

Key changes include:
- A new `features` directory has been created to house all business-related modules.
- Components from the legacy [pages](cci:7://file:///c:/Users/SL310/Documents/html/electron/AquaNote/src/app/pages:0:0-0:0) and [core](cci:7://file:///c:/Users/SL310/Documents/html/electron/AquaNote/src/app/core:0:0-0:0) directories (e.g., Home, Login, Register) have been migrated into the `features` directory.
- Feature-specific services and data models are now co-located with their corresponding feature modules (e.g., `LoginService` is now inside [features/login](cci:7://file:///c:/Users/SL310/Documents/html/electron/AquaNote/src/app/features/login:0:0-0:0)).
- All import paths across the application have been updated to reflect the new file locations.
- The now-empty [pages](cci:7://file:///c:/Users/SL310/Documents/html/electron/AquaNote/src/app/pages:0:0-0:0) and [core/models](cci:7://file:///c:/Users/SL310/Documents/html/electron/AquaNote/src/app/core/models:0:0-0:0) directories have been removed.
